requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asrpy=0.0.3=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
autoreject=0.4.3=pypi_0
blas=2.131=openblas
blas-devel=3.9.0=31_h1ea3ea9_openblas
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
coloredlogs=15.0.1=pyhd8ed1ab_4
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.9.19=0
cuda-runtime=12.4.1=0
cuda-version=12.9=3
cudnn=9.10.1.4=h2b6041c_0
cycler=0.12.1=pypi_0
darkdetect=0.8.0=pyhd8ed1ab_1
debugpy=1.8.11=py312h6a678d5_0
decorator=5.2.1=pyhd8ed1ab_0
deprecated=1.2.13=py312h06a4308_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.0=h5888daf_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
fonttools=4.57.0=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2024.6.1=pypi_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gnutls=3.6.15=he1e5248_0
h5io=0.2.5=pypi_0
h5py=3.13.0=pypi_0
humanfriendly=10.0=pyh707e725_8
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=9.2.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
joblib=1.5.0=pypi_0
jpeg=9e=h5eee18b_3
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=pypi_0
lame=3.100=h7b6447c_0
lazy_loader=0.4=py312h06a4308_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libblas=3.9.0=31_h59b9bed_openblas
libcblas=3.9.0=31_he106b2a_openblas
libcublas=12.4.5.8=0
libcudnn=9.10.1.4=h24322f0_0
libcudnn-dev=9.10.1.4=h2b6041c_0
libcufft=11.2.1.3=0
libcufile=1.14.0.30=4
libcurand=10.3.10.19=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=h5eee18b_0
libexpat=2.7.0=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=15.1.0=h767d61c_2
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=31_h7ac8fdf_openblas
liblapacke=3.9.0=31_he2f377e_openblas
libnpp=12.2.5.30=0
libnsl=2.0.1=hd590300_0
libnvfatbin=12.9.19=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libopenblas=0.3.29=ha39b09d_0
libpng=1.6.37=h21135ba_2
libprotobuf=5.29.3=hc99497a_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libtorch=2.5.1=gpu_cuda124_he07d216_204
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.8=hfdd30dd_0
libzlib=1.2.13=h4ab18f5_6
llvmlite=0.44.0=py312hc1e8f15_1
lz4-c=1.9.3=h9c3ff4c_1
magma=2.7.1=ha545f22_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-base=3.10.0=py312hbfdbfaf_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
meegkit=0.1.9=pypi_0
mkl=2023.1.0=h213fc3f_46344
mne-base=1.9.0=pyh694c41f_201
mne-icalabel=0.7.0=pyhd8ed1ab_1
mne-qt-browser=0.7.1=pyh0555025_0
mpmath=1.3.0=pyhd8ed1ab_1
nccl=2.21.5.1=ha515578_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
nettle=3.7.3=hbbd107a_1
networkx=3.3=pypi_0
nomkl=3.0=0
numba=0.61.2=py312h2e6246c_0
numexpr=2.10.1=py312hd28fd6d_0
numpy=2.2.5=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.1.105=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
ocl-icd=2.3.2=h5eee18b_1
onnxruntime=1.20.1=py312hce09735_0
openblas=0.3.29=h06a4308_0
openblas-devel=0.3.29=ha39b09d_0
openh264=2.1.1=h4ff587b_0
openssl=3.5.0=h7b32b05_1
opentelemetry-api=1.30.0=py312h06a4308_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pyhe01879c_0
pooch=1.8.2=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
protobuf=5.29.3=py312he621ea3_0
psutil=5.9.0=py312h5eee18b_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.19.1=pyhd8ed1ab_0
pymatreader=1.0.0=pypi_0
pynndescent=0.5.13=pyhd8ed1ab_1
pyopengl=3.1.9=pyhd8ed1ab_0
pyparsing=3.2.3=pypi_0
pyqtgraph=0.13.7=pyhd8ed1ab_1
pyriemann=0.8=pypi_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-flatbuffers=25.2.10=pyhbc23db3_0
python-picard=0.8=pyha308f57_1
python_abi=3.12=7_cp312
pytorch=2.5.1=gpu_cuda124_py312h307a77b_204
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pywavelets=1.8.0=py312h5eee18b_0
pyzmq=26.2.0=py312h6a678d5_0
qdarkstyle=3.2.3=pyhd8ed1ab_1
qtpy=2.4.3=pyhd8ed1ab_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.3=pypi_0
scooby=0.10.1=pyhd8ed1ab_0
setuptools=78.1.1=py312h06a4308_0
six=1.17.0=pypi_0
sleef=3.5.1=h5eee18b_2
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h39e8969_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1=py312_cu124
torchvision=0.20.1=py312_cu124
tornado=6.4.2=py312h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
triton=3.1.0=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
umap-learn=0.5.7=py312h7900ff3_1
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py312h06a4308_0
wrapt=1.17.0=py312h5eee18b_0
xmltodict=0.14.2=pypi_0
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h4ab18f5_6
zstd=1.5.2=h8a70e8d_1
